refactor renter/host/session
