+++ core/commands/storage/renter/storage.go
@@ -2,7 +2,6 @@ package renter
 
 import (
 	"context"
-	"encoding/hex"
 	"fmt"
 	"math"
 	"strconv"
@@ -10,8 +9,8 @@ import (
 
 	"github.com/TRON-US/go-btfs/core/commands/cmdenv"
 	"github.com/TRON-US/go-btfs/core/commands/storage"
-	"github.com/TRON-US/go-btfs/core/commands/storage/guard"
 	"github.com/TRON-US/go-btfs/core/commands/storage/host"
+	"github.com/TRON-US/go-btfs/core/commands/storage/renter/guard"
 	"github.com/TRON-US/go-btfs/core/corehttp/remote"
 	"github.com/TRON-US/go-btfs/core/escrow"
 	"github.com/TRON-US/go-btfs/core/hub"
